Delete python script
